move `BatchedGraph`
