+++ mol_gnn/data/__init__.py
@@ -1,4 +1,5 @@
-from .batch import BatchedGraph, MpnnBatch, MultiInputMpnnBatch
+from .models.graph import BatchedGraph
+from .batch import MpnnBatch, MultiInputMpnnBatch
 from .mixins import Datum
 from .molecule import MoleculeDatapoint, MoleculeDataset
 from .multi import MultiInputDataset
