+++ mol_gnn/data/mixins.py
@@ -9,9 +9,9 @@ import torch
 from torch.utils.data import DataLoader
 
 from mol_gnn.types import Mol
-from mol_gnn.data.batch import BatchedGraph, MpnnBatch
+from mol_gnn.data.batch import MpnnBatch
 from mol_gnn.featurizers.vector.base import VectorFeaturizer
-from mol_gnn.data.models.graph import Graph
+from mol_gnn.data.models.graph import BatchedGraph, Graph
 
 
 @dataclass(slots=True)
