+++ mol_gnn/models/multi.py
@@ -3,7 +3,7 @@ from typing import Iterable
 import torch
 from torch import Tensor
 
-from mol_gnn.data import BatchedGraph
+from mol_gnn.data.models.graph import BatchedGraph
 from mol_gnn.nn import MultiInputMessagePassing, Aggregation, Predictor
 from mol_gnn.models.mpnn import MPNN
 from mol_gnn.nn.metrics import Metric
